requirements.txt
openai
pandas
numpy
flask
psutil
fastmcp
win32api
win32con
numpy
os
